This PR just updates some dyno-tests that were making comparisons
between `int` and `size_t` types, which caused a warning that gets
promoted to an error on some systems (Ubuntu, for example).

It also addresses a warning about missing curly braces in an array
initializer in one test.

[reviewed by @dlongnecke-cray]

+++ frontend/test/resolution/testSubtypePrimitives.cpp
@@ -68,7 +68,7 @@ static void testPrimitive(std::string primitive,
 
   auto varTypes = resolveTypesOfVariables(context, ps.str(), variables);
 
-  for (int i = 0; i < varTypes.size(); i++) {
+  for (size_t i = 0; i < varTypes.size(); i++) {
     std::cout << "Checking " << variables[i] << std::endl;
 
     if (expectedResults[i] == shouldReturnTrue) {
